[MODEXPW-529] Move tests to appropriate packages
