+++ src/test/java/org/folio/dew/batch/acquisitions/edifact/services/HoldingServiceTest.java
@@ -1,9 +1,8 @@
-package org.folio.dew.batch.acquisitions.edifact;
+package org.folio.dew.batch.acquisitions.edifact.services;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import org.folio.dew.batch.acquisitions.edifact.services.HoldingService;
 import org.folio.dew.client.HoldingClient;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
