+++ src/test/java/org/folio/dew/batch/acquisitions/edifact/services/IdentifierTypeServiceTest.java
@@ -1,10 +1,9 @@
-package org.folio.dew.batch.acquisitions.edifact;
+package org.folio.dew.batch.acquisitions.edifact.services;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
-import org.folio.dew.batch.acquisitions.edifact.services.IdentifierTypeService;
 import org.folio.dew.client.IdentifierTypeClient;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
