+++ src/test/java/org/folio/dew/batch/acquisitions/edifact/services/MaterialTypeServiceTest.java
@@ -1,9 +1,8 @@
-package org.folio.dew.batch.acquisitions.edifact;
+package org.folio.dew.batch.acquisitions.edifact.services;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.folio.dew.BaseBatchTest;
-import org.folio.dew.batch.acquisitions.edifact.services.MaterialTypeService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
